+++ Opener/Models/OKey.cs
@@ -6,7 +6,7 @@ public class OKey
 {
     public string Id { get; set; } = Guid.NewGuid().ToString();
     public string Key { get; set; } = string.Empty;
-    public OKeyType KeyType { get; set; }
+    public OKeyType KeyType { get; set; } = OKeyType.Data;
     public string Value { get; set; } = string.Empty; // Renamed from Path for clarity across types
     public string Description { get; set; } = string.Empty;
 }